fixed requests to Ensembl and tests
